Moved to `semantic_scholar` package
