+++ sciety_labs/providers/opensearch/async_providers.py
@@ -16,7 +16,7 @@ from sciety_labs.providers.interfaces.async_article_recommendation import (
     AsyncSingleArticleRecommendationProvider
 )
 from sciety_labs.providers.async_crossref import AsyncCrossrefMetaDataProvider
-from sciety_labs.providers.async_semantic_scholar import (
+from sciety_labs.providers.semantic_scholar.async_semantic_scholar import (
     AsyncSemanticScholarTitleAbstractEmbeddingVectorProvider
 )
 from sciety_labs.providers.opensearch.utils import (
